+++ src/components/common/Header.js
@@ -6,10 +6,11 @@ import { Button, Container, Dropdown, Image, Nav, Navbar } from "react-bootstrap
 
 import { isLogin } from "../../utils/TokenUtils";
 import { reset } from "../../modules/UserModules";
-import { callLogoutAPI, callProfileAPI } from "../../apis/UserAPICalls";
+import { callLogoutAPI } from "../../apis/AuthAPICalls";
+import { callProfileAPI } from "../../apis/UserAPICalls";
 
 import ChatBox from "../item/ChatBox";
-import { DefaultProfile } from "./Icons";
+import ProfileImage from "./ProfileImage";
 
 const CustomNavLink = ({ to, children }) => (
     <NavLink
@@ -79,15 +80,8 @@ function Header() {
                             <Dropdown.Menu className="dropdown-animation dropdown-menu-end shadow" aria-labelledby="profileDropdown">
                                 <Dropdown.Item as="div">
                                     <div className="d-flex align-items-center">
-                                        <div className=" me-3 rounded-5 overflow-hidden border">
-                                            {!profileInfo ?
-                                                (
-                                                    <DefaultProfile
-                                                        className="avatar mx-auto d-block mb-3"
-                                                        width="50px" height="50px"
-                                                    />)
-                                                : <Image src={profileInfo?.profile.profileImage} width="50px" height="50px" />
-                                            }
+                                        <div className="me-3">
+                                            <ProfileImage />
                                         </div>
                                         <div>
                                             <p className="small m-0">
